pyproject.toml
[project]
name = "dunky"
version = "0.1.0"
description = "A Jupyter Kernel for DuckDB with Unity Catalog"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbt-duckdb[unity]",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "pandas==2.2.2",
    "tabulate==0.9.0",
    "jupyter-client>=8.6.2",
]

[tool.uv]
dev-dependencies = [
    "coverage>=7.6.1",
    "hatchling>=1.25.0",
    "pre-commit>=3.8.0",
    "pytest>=8.3.3",
    "ruff>=0.6.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.hooks.custom]
path = "build.py"

[tool.uv.sources]
dbt-duckdb = { git = "https://github.com/godatadriven/dbt-duckdb.git" }

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "W", "C90"]
lint.ignore = ["E501"]
lint.per-file-ignores = {"__init__.py" = ["F401", "F403"]}
exclude = ["build", "dist", "venv"]
